Fixing unsigned error
git-svn-id: http://calder.sdml.cs.kent.edu/svn/srcDiffPurged/trunk/srcdiff/trunk@9364 f70b1346-9fe7-0310-9bf3-95a0de396e6b

+++ src/srcdiff.cpp
@@ -761,12 +761,12 @@ void output_diffs(struct reader_buffer * rbuf_old, std::vector<std::vector<int> 
 
     diff_end_old = rbuf_old->last_output;
     diff_end_new = rbuf_new->last_output;
-    if(edits->operation == DELETE && last_diff_old < (unsigned)edits->offset_sequence_one) {
+    if(edits->operation == DELETE && last_diff_old < edits->offset_sequence_one) {
 
       diff_end_old = node_sets_old->at(edits->offset_sequence_one - 1)->back() + 1;
       diff_end_new = node_sets_new->at(last_diff_new + (edits->offset_sequence_one - last_diff_old) - 1)->back() + 1;
 
-    } else if(edits->operation == INSERT && last_diff_old <= (unsigned)edits->offset_sequence_one) {
+    } else if(edits->operation == INSERT && last_diff_old <= edits->offset_sequence_one) {
 
       diff_end_old = node_sets_old->at(edits->offset_sequence_one)->back() + 1;
       diff_end_new = node_sets_new->at(last_diff_new + (edits->offset_sequence_one - last_diff_old))->back() + 1;
@@ -962,12 +962,12 @@ void output_comment_paragraph(struct reader_buffer * rbuf_old, std::vector<std::
     // add preceeding unchanged
     diff_end_old = rbuf_old->last_output;
     diff_end_new = rbuf_new->last_output;
-    if(edits->operation == DELETE && last_diff_old < (unsigned)edits->offset_sequence_one) {
+    if(edits->operation == DELETE && last_diff_old < edits->offset_sequence_one) {
 
       diff_end_old = node_sets_old->at(edits->offset_sequence_one - 1)->back() + 1;
       diff_end_new = node_sets_new->at(last_diff_new + (edits->offset_sequence_one - last_diff_old) - 1)->back() + 1;
 
-    } else if(edits->operation == INSERT && last_diff_old <= (unsigned)edits->offset_sequence_one) {
+    } else if(edits->operation == INSERT && last_diff_old <= edits->offset_sequence_one) {
 
       diff_end_old = node_sets_old->at(edits->offset_sequence_one)->back() + 1;
       diff_end_new = node_sets_new->at(last_diff_new + (edits->offset_sequence_one - last_diff_old))->back() + 1;
@@ -1079,12 +1079,12 @@ void output_comment_line(struct reader_buffer * rbuf_old, std::vector<std::vecto
     // add preceeding unchanged
     diff_end_old = rbuf_old->last_output;
     diff_end_new = rbuf_new->last_output;
-    if(edits->operation == DELETE && last_diff_old < (unsigned)edits->offset_sequence_one) {
+    if(edits->operation == DELETE && last_diff_old < edits->offset_sequence_one) {
 
       diff_end_old = node_sets_old->at(edits->offset_sequence_one - 1)->back() + 1;
       diff_end_new = node_sets_new->at(last_diff_new + (edits->offset_sequence_one - last_diff_old) - 1)->back() + 1;
 
-    } else if(edits->operation == INSERT && last_diff_old <= (unsigned)edits->offset_sequence_one) {
+    } else if(edits->operation == INSERT && last_diff_old <= edits->offset_sequence_one) {
 
       diff_end_old = node_sets_old->at(edits->offset_sequence_one)->back() + 1;
       diff_end_new = node_sets_new->at(last_diff_new + (edits->offset_sequence_one - last_diff_old))->back() + 1;
